refactor(search): extract map-fly-to, search-dates, price-input into src/lib/search
Slice 1 of the unified SearchBar plan. MAP_FLY_TO_EVENT moves to
src/lib/search/map-fly-to.ts (SearchForm re-exports for back-compat until
it is retired); the triplicated validateMoveInDate gains a canonical home
in search-dates.ts (SearchForm + DesktopHeaderSearch deduped; InlineFilterStrip
left untouched deliberately); DesktopHeaderSearch's price clamp/swap logic is
codified as normalizePriceRange in price-input.ts for the shared submit
pipeline to come. No behavior change.
